Added goal for calling Soy header compiler.
While this goal works, for it to be useful we need to be able to control the directory from which the header compiler runs. Otherwise there is no way for us to specify a source like `nl/serviceplanet/valygar/common/web/soy/templates/base-components.soy` to generate a header file for. And we need to be able to specify a source like that because the path ends up in the header file.

Ref T11502.