Allow Soy header compiler to work with relative paths
Summary:
The Soy header compiler uses the full path name provided in the `--srcs` as the name for the template in the header file. Because this can't be an absolute path we need to be able to supply a path relative to `${project.dir}/src/main/resources`.

Ref T11502.

Reviewers: siepkes

Reviewed By: siepkes

Maniphest Tasks: T11502

Differential Revision: https://phabricator.serviceplanet.nl/D6636